mm/damon/vaddr: support perf event based access check
This patch adds perf event based access check for virtual address space
monitoring.

The perf event that can be specified for perf event-based access check
must be able to obtain the data source address corresponding to the
sample.  In other words, PERF_SAMPLE_DATA_SRC must be able to be
specified in perf_event_attr.sample_type

The perf event based access check is performed as follows:

1. During the sampling interval, samples of the perf event are stored in
   a sampling buffer.

2. After the sampling interval ends, during the access check, a
   histogram showing the number of accesses to each memory address is
   generated based on each sample contained in the sampling buffer.

If the sampling buffer is full during a sampling interval, the sampling
buffer is expanded to prepare for the next sampling interval.

If you configure the memory addresses corresponding to the perf event
samples to be obtained as virtual addresses, the histogram is
implemented as a two-dimensional xarray indexed by pid and virtual
memory address.

If you configure perf event samples to have their corresponding memory
addresses retrieved by physical addresses, the histogram is implemented
using an xarray indexed by physical addresses.

3. For each monitoring target region, the number of accesses is updated
   based on the histogram.

Signed-off-by: Akinobu Mita <akinobu.mita@gmail.com>