cxl/hmu: allow hotlist sampling from perf event overflow handler
This patch adds the CXL Hotness Monitoring Unit perf driver [1].

The original driver allows the hot list to be retrieved from the perf aux buffer.
However, DAMON's perf event-based access check does not support sampling
from the aux buffer, so this patch also extends it to be able to sample using
the overflow handler specified when initializing the perf event with
perf_event_create_kernel_counter().

Currently, the node number corresponding to the device is hard-coded, so
this needs to be fixed.

[1] https://lore.kernel.org/lkml/20241121101845.1815660-1-Jonathan.Cameron@huawei.com/